Add sniffer runner to find all unique events
Added a custom runner that parses log files and collects unique lines.
The purpose of this is mostly to look for new events and variations of
known events to get a better idea of how each event payload must look.

To make this work I had to break out the mapping logic out of
log-parser.ts, this allows us to reuse all the splitting code without
tying it into the event-mapper code.